pyproject.toml
[project]
name = "fastapi-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "joblib>=1.5.2",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

#=은 버전 그대로 , >=은 버전 이상의 것을 의존성 추가(메이저 버전은 포함하지 않음)
# dockerfile처럼 협업 파트너간의 환경 통일을 위해 사용

